// src/types.ts
  export interface CarTelemetryData {
    m_speed: number;                              // Speed of car in kilometres per hour
    m_throttle: number;                           // Amount of throttle applied (0.0 to 1.0)
    m_steer: number;                              // Steering (-1.0 (full lock left) to 1.0 (full lock right))
    m_brake: number;                              // Amount of brake applied (0.0 to 1.0)
    m_clutch: number;                              // Amount of clutch applied (0 to 100)
    m_gear: number;                                // Gear selected (1-8, N=0, R=-1)
    m_engineRPM: number;                          // Engine RPM
    m_drs: number;                                 // 0 = off, 1 = on
    m_revLightsPercent: number;                    // Rev lights indicator (percentage)
    m_revLightsBitValue: number;                  // Rev lights (bit 0 = leftmost LED, bit 14 = rightmost LED)
    m_brakesTemperature: number[];             // Brakes temperature (celsius)
    m_tyresSurfaceTemperature: number[];        // Tyres surface temperature (celsius)
    m_tyresInnerTemperature: number[];          // Tyres inner temperature (celsius)
    m_engineTemperature: number;                  // Engine temperature (celsius)
    m_tyresPressure: number[];                 // Tyres pressure (PSI)
    m_surfaceType: number[];                    // Driving surface, see appendices
  }
  
  export interface PacketCarTelemetryData {
    m_header: PacketHeader;                                      // Header
    m_carTelemetryData: CarTelemetryData[];
    m_mfdPanelIndex: number;                                         // Index of MFD panel open - 255 = MFD closed
  
    m_mfdPanelIndexSecondaryPlayer: number;                          // See above
    m_suggestedGear: number;                                         // Suggested gear for the player (1-8)
    // 0 if no gear suggested
  }
// ...
  export interface PacketHeader {
    packet_format: number;           // u16
    game_year: number;               // u8
    game_major_version: number;      // u8
    game_minor_version: number;      // u8
    packet_version: number;          // u8
    packet_id: number;               // u8
    session_uid: bigint;             // u64, BigInt in JavaScript/TypeScript/ convert to string for use with JSON
    session_time: number;            // f32
    frame_identifier: number;        // u32
    overall_frame_identifier: number;// u32
    player_car_index: number;        // u8
    secondary_player_car_index: number; // u8
  }